expose isExternal, isBuildable can be to narrow
I need to know whether a project comes from one of the directories I
analyse or was loaded through aether. isBuildable is to narrow in the
sense that a particular project isExternal but does not have a src
folder (e.g. a parent pom without any src) in which case isBuildable
returns false.