return Gav for getSubmodules
returning Project is not ideal because we either have to filter projects
which are not part of the repository (so it might be an incomplete
result) or we would have to throw an exception which is not what I need.
Thus return Gavs instead, the user can then retrieve Projects via
Session if needed.